fix(workspace): move extracted module tests to conventional _test.go files (#654)
- Create stream_processor_test.go: TestStreamProcessor_appendErr + processLine edge cases
- Create pipeline_test.go: WirePipes cleanup tests + StartFailure + captureStderrAsync
- Create confluence_format_test.go: xhtmlToMarkdown + markdownToXhtml table tests
- Remove moved tests from original test files
