+++ internal/tools/workspace/process_executor_unit_test.go
@@ -225,122 +225,6 @@ func TestProcessExecutor_LookPath(t *testing.T) {
 }
 
 // ---------------------------------------------------------------------------
-// streamProcessor.appendErr tests
-// ---------------------------------------------------------------------------
-
-// verifyAppendErrResult asserts the state of the string builder, truncated flag,
-// and optional feedback buffer against the test case expectations.
-func verifyAppendErrResult(t *testing.T, sb *strings.Builder, sp *streamProcessor, wantEmptySB bool, wantContains string, wantTruncated bool, useFeedback bool, fb io.Writer) {
-	t.Helper()
-	got := sb.String()
-	if wantEmptySB && got != "" {
-		t.Errorf("expected empty sb, got %q", got)
-	}
-	if wantContains != "" && !strings.Contains(got, wantContains) {
-		t.Errorf("expected sb to contain %q, got %q", wantContains, got)
-	}
-	if sp.truncated.Load() != wantTruncated {
-		t.Errorf("truncated = %v, want %v", sp.truncated.Load(), wantTruncated)
-	}
-	if useFeedback && fb != nil {
-		fbStr := fb.(*bytes.Buffer).String()
-		if !strings.Contains(fbStr, wantContains) {
-			t.Errorf("expected feedback to contain %q, got %q", wantContains, fbStr)
-		}
-	}
-}
-
-func TestStreamProcessor_appendErr(t *testing.T) {
-	tests := []struct {
-		name          string
-		totalCaptured int
-		maxCapture    int
-		err           error
-		wantContains  string
-		wantTruncated bool
-		wantEmptySB   bool
-		useFeedback   bool
-	}{
-		{
-			name:          "nil error — no-op",
-			totalCaptured: 0,
-			maxCapture:    500,
-			err:           nil,
-			wantContains:  "",
-			wantTruncated: false,
-			wantEmptySB:   true,
-		},
-		{
-			name:          "ErrTooLong — warning without error detail",
-			totalCaptured: 0,
-			maxCapture:    500,
-			err:           bufio.ErrTooLong,
-			wantContains:  "[Warning] Output line too long",
-			wantTruncated: false,
-		},
-		{
-			name:          "generic error — warning with error text",
-			totalCaptured: 0,
-			maxCapture:    500,
-			err:           fmt.Errorf("mock read error"),
-			wantContains:  "[Warning] Output read error: mock read error",
-			wantTruncated: false,
-		},
-		{
-			name:          "truncated by remaining capacity",
-			totalCaptured: 490,
-			maxCapture:    500,
-			err:           fmt.Errorf("some error"),
-			wantContains:  "\n[Warning]",
-			wantTruncated: true,
-			// Only 10 remaining bytes; the full warning + newline exceeds it,
-			// so truncated.Store(true) is set AND truncateToValidUTF8
-			// cuts to the first 10 valid UTF-8 bytes: "\n[Warning]".
-		},
-		{
-			name:          "already at max — truncated flag set, nothing written",
-			totalCaptured: 500,
-			maxCapture:    500,
-			err:           fmt.Errorf("some error"),
-			wantContains:  "",
-			wantTruncated: true,
-			wantEmptySB:   true,
-		},
-		{
-			name:          "error with feedback writer",
-			totalCaptured: 0,
-			maxCapture:    500,
-			err:           fmt.Errorf("feedback test error"),
-			wantContains:  "[Warning] Output read error: feedback test error",
-			wantTruncated: false,
-			useFeedback:   true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			totalCaptured := tt.totalCaptured
-			var fb io.Writer
-			if tt.useFeedback {
-				fb = &bytes.Buffer{}
-			}
-			sp := &streamProcessor{
-				mu:            &sync.Mutex{},
-				truncated:     &atomic.Bool{},
-				totalCaptured: &totalCaptured,
-				maxCapture:    tt.maxCapture,
-				feedback:      fb,
-			}
-
-			var sb strings.Builder
-			sp.appendErr(&sb, tt.err)
-
-			verifyAppendErrResult(t, &sb, sp, tt.wantEmptySB, tt.wantContains, tt.wantTruncated, tt.useFeedback, fb)
-		})
-	}
-}
-
-// ---------------------------------------------------------------------------
 // writeTracker.Write tests
 // ---------------------------------------------------------------------------
